msmnile: Set Header version for bootimage
Set Header version to 1 for bootimage.

Change-Id: Idd2c78689f292dbcdfa7f1eb0b91a8d91aac12c6
Signed-off-by: Lijuan Gao <lijuang@codeaurora.org>

msmnile: Add support for header version 2 of boot image

Adding support to make a bootimage with header version as 2.
This bootimage no longer has the dtbs appended to the kernel.
All the dtbs are made into a new image called dtb.img and
this image is present in the bootimage.

Change-Id: I47867fc284a45bcf2e54af5f6cd40d894418c002
Signed-off-by: Lijuan Gao <lijuang@codeaurora.org>